refactor: 당첨번호와 보너스 번호를 따로 관리하도록 구조 변경
